so far the same as main
for limited/minimum html - trying to get a dropdown's input to affect which element gets outputed (meaning you'd have a bunch of elements variables in what is now called elements_orig.js [cy-conf/elements.js in this branch] and choosing which one to export based on input from the dropdown). You'd only have one index.html file and you'd change the elements that get exported in cy-conf/elements.js. The purpose is to try a method that is less data-heavy